Remove 1st move "only 1 quadrant" search
The first computer move chose only 1 quadrant to
look for a move, since the board is rotationally
symmetric at that point. This ended up biasing the
rest of the game in a bad way, because the tree of
struct Node gets carried through to further moves in
the game. Really bad idea.